requirements.txt
langchain==0.3.7
langchain-openai==0.2.9
langchain-community==0.3.7
langchain-text-splitters==0.3.3
langchain-chroma==0.1.4
chromadb==0.5.23
openai==1.55.3
sqlalchemy==2.0.35
psycopg[binary]>=3.1.0
rank-bm25==0.2.2
python-dotenv==1.0.1
